Integrate mutlizone, display (#40)
* Log multizone and side tofs together

* Add screen

* Improve start procedure for display

* Show left/right tofs on logger: incorrect

* Cleanup

* Partial re-write of side tof library

* Show start screen for 2 seconds

* Driving animation

* Cleanup

* Cleanup

* Cleanup display

* Cleanup unused methods

* Variable standardization

---------

+++ src/actuators/rgb_led.cpp
@@ -25,7 +25,7 @@ void RgbLed::set_led_white() {
     set_led_to_color(255, 255, 255);
 }
 
-void RgbLed::set_led_to_color(int red, int green, int blue) {
+void RgbLed::set_led_to_color(uint8_t red, uint8_t green, uint8_t blue) {
     strip.setPixelColor(0, strip.Color(red, green, blue));
     strip.show();
 }
