+++ src/lab_demo/lab_demo_manager.h
@@ -28,5 +28,5 @@ class LabDemoManager : public Singleton<LabDemoManager> {
         int16_t nextRightSpeed;
 
         // Constants
-        static constexpr int MIN_ENCODER_PULSES = 10;
+        static constexpr uint8_t MIN_ENCODER_PULSES = 10;
 };
